More GeoModel cleanup.
Changed `getGeometry()` to return all geometry objects if no type is
specified explicitly (previously, it returned just Polygon objects).
This seems like a better default behavior.

Refactored `cleanPolygon()` and `cleanGeometryCoordinates()` so all
Polygon cleanup is in one place.

Refactored tests to split out GeoModel tests into a separate
`GeoModel.js` test file. Added a bunch of GeoModel tests.
